User registration works git add . DB needs alterations for storing hashes
